+++ R/data.R
@@ -190,11 +190,11 @@
 #' of these questions as accurately as possible.", data on n = 20174
 #' participants remains.
 #'
-#' @section Items
+#' @section Items:
 #'
 #'   \url{https://ipip.ori.org/newhexaco_pi_key.htm}
 #'
-#' @section source
+#' @section source:
 #'
 #'   \url{https://openpsychometrics.org/_rawdata/} (May 6th, 2020)
 "HEXACO"